refactor!: store money as integer minor units & centralize money math
Persist all monetary columns as integer cents (Prisma Int) and route every
money operation through a single shared Money value object, replacing four
divergent toFixed/roundToCents implementations across pricing, coupons, the
70% booster split & the wallet ledger.

- add @packages/shared/money/money value object (cents canonical, half-up)
- migrate Order/Payment/Wallet/WalletTransaction/PricingStep/extras Float->Int
  via ROUND(value * 100); keep reputation, modifierRate & dual-semantic
  Coupon.discount as Float
- convert cents->reais only at human/provider boundaries: Mercado Pago
  unit_price, BRL formatters & the withdrawal input
- express @packages/config price tables in cents; require integer-cent inputs
  on admin pricing & wallet request schemas

Implements the money slice (#71a) of the hard technical-debt epic #71.

BREAKING CHANGE: monetary database columns and API money fields now use
integer minor units (cents) instead of floating-point major units (reais).
A data-converting Prisma migration is included; clients must format cents.

+++ apps/api/src/modules/wallet/presentation/wallets.request-schemas.ts
@@ -15,7 +15,7 @@ export type ListWalletTransactionsQuerySchemaInput = z.infer<
 export const creditCompletedOrderEarningsSchema = z.object({
 	orderId: z.string().trim().min(1),
 	boosterId: z.string().trim().min(1),
-	amount: z.number().positive(),
+	amount: z.number().int().positive(),
 	completedAt: z.string().datetime(),
 	lockPeriodInHours: z.number().int().nonnegative(),
 });
@@ -35,7 +35,7 @@ export type ReleaseMaturedWalletFundsSchemaInput = z.infer<
 >;
 
 export const requestWithdrawalSchema = z.object({
-	amount: z.number().positive(),
+	amount: z.number().int().positive(),
 	requestedAt: z.string().datetime(),
 });
